Move CORS caches to tenant space
